+++ topobank/manager/admin.py
@@ -1,6 +1,11 @@
 from django.contrib import admin
 
-from .models import Surface, Topography
+from .models import Property, Surface, Topography
+
+
+@admin.register(Property)
+class PropertyAdmin(admin.ModelAdmin):
+    list_display = ('id', 'name', 'value_categorical', 'value_numerical', 'unit')
 
 
 @admin.register(Surface)