+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.7.0 (not yet released)
 
+- ENH: Properties as key-value pairs with categorical and numerical values,
+  including units
 - ENH: Tags as analysis subjects (as a replacement for surface collections)
 - ENH: Import of ZIP containers from URL